+++ client/src/components/Auth.jsx
@@ -34,7 +34,7 @@ const Auth = ({mode}) => {
     return (
         <div className='formContainer'>
             <div className='formWrapper'>
-                <span className='logo'>Attendance</span>
+                <span className='logo'>SmartWork</span>
                 <span className='title'>{mode}</span>
                 <form onSubmit={handleSubmit}>
                     {!(mode==='login')&&(<input id='name' type='text' placeholder='name' onChange={handleChange}/>)}