+++ client/src/pages/Register.jsx
@@ -1,19 +1,15 @@
-import Sidebar from '../components/Sidebar'
 import Navbar from '../components/Navbar'
 import Auth from '../components/Auth'
 import Footer from '../components/Footer'
 
 const Register = () => {   
     return (
-        <div className='container'>
-            <Sidebar/>
-            <div className='wrapper'>
-                <Navbar/> 
-                <Auth
-                    mode='register'
-                />
-                <Footer/>
-            </div>
+        <div>
+            <Navbar/> 
+            <Auth
+                mode='register'
+            />
+            <Footer/>
         </div>
     )
 }